+++ scripts/tools/note.js
@@ -436,3 +436,74 @@ function checkEndNote(note1, note2) {
         return true;
     return false;
 }
+
+function detectShrAngle(notes, bpm, offset, bpmc) {
+    let arr = [];
+    let lastRed;
+    let lastRedDir;
+    let startRedDot;
+    let lastBlue;
+    let lastBlueDir;
+    let startBlueDot;
+    for (let i = 0, len = notes.length; i < len; i++) {
+        const note = notes[i];
+        if (note._type == 0) {
+            if (lastRed) {
+                if ((maybeWindowed(note, lastRed) && checkTolerance(note, lastRed, maxWindowTolerance)) || checkTolerance(note, lastRed, maxTolerance)) {
+                    if (startRedDot) {
+                        startRedDot = null;
+                        lastRedDir = flipCutDir[lastRedDir];
+                    }
+                    if (checkShrAngle(note._cutDirection, lastRedDir) && toRealTime(note._time - lastRed._time) < shrAngleMax / 1000) {
+                        arr.push(adjustTime(note._time, bpm, offset, bpmc));
+                    }
+                    if (note._cutDirection == 8) startRedDot = note;
+                    else lastRedDir = note._cutDirection;
+                }
+                else {
+                    if (startRedDot && checkShrAngle(note._cutDirection, lastRedDir) && toRealTime(note._time - lastRed._time) < shrAngleMax / 1000) {
+                        arr.push(adjustTime(startRedDot._time, bpm, offset, bpmc));
+                        startRedDot = null;
+                    }
+                    if (note._cutDirection != 8) lastRedDir = note._cutDirection;
+                }
+            }
+            else lastRedDir = note._cutDirection;
+            lastRed = note;
+        }
+        else if (note._type == 1) {
+            if (lastBlue) {
+                if ((maybeWindowed(note, lastBlue) && checkTolerance(note, lastBlue, maxWindowTolerance)) || checkTolerance(note, lastBlue, maxTolerance)) {
+                    if (startBlueDot) {
+                        startBlueDot = null;
+                        lastBlueDir = flipCutDir[lastBlueDir];
+                    }
+                    if (checkShrAngle(note._cutDirection, lastBlueDir) && toRealTime(note._time - lastBlue._time) < shrAngleMax / 1000) {
+                        arr.push(adjustTime(note._time, bpm, offset, bpmc));
+                    }
+                    if (note._cutDirection == 8) startBlueDot = note;
+                    else lastBlueDir = note._cutDirection;
+                }
+                else {
+                    if (startBlueDot && checkShrAngle(note._cutDirection, lastBlueDir) && toRealTime(note._time - lastBlue._time) < shrAngleMax / 1000) {
+                        arr.push(adjustTime(startBlueDot._time, bpm, offset, bpmc));
+                        startBlueDot = null;
+                    }
+                    if (note._cutDirection != 8) lastBlueDir = note._cutDirection;
+                }
+            }
+            else lastBlueDir = note._cutDirection;
+            lastBlue = note;
+        }
+    }
+    arr = arr.filter(function(x, i, ary) {
+        return !i || x != ary[i - 1];
+    });
+    return arr;
+}
+
+function checkShrAngle(n1cd, n2cd) {
+    if (n1cd == 8 || n2cd == 8) return false;
+    if ((n2cd == 6 || n2cd == 7) && n1cd == 0) return true;
+    return false;
+}